Move hook into separate module
